rename get-profile action to follow project convention
